# backend/docs/swagger.yaml
basePath: /v1/api/radar-hub-manager
definitions:
...
  github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Station:
    properties:
      created_at:
        type: integer
      distance_to_coast:
        description: km
        type: number
      elevation:
        description: mét so với mực nước biển
        type: number
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      note:
        description: ghi chú tự do
        type: string
      status:
        description: '"ACTIVE" / "INACTIVE" (tính từ lịch)'
        type: string
      updated_at:
        type: integer
    type: object
...
  internal_handlers.CreateStationRequest:
    properties:
      distance_to_coast:
        example: 15.2
        type: number
      elevation:
        example: 10.5
        type: number
      id:
        example: 1
        type: integer
      latitude:
        example: 21.0285
        type: number
      longitude:
        example: 105.8542
        type: number
      name:
        example: Station Alpha
        type: string
      note:
        example: Main radar station
        type: string
      status:
        example: ACTIVE
        type: string
    required:
    - latitude
    - longitude
    - name
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_handlers.UpdateStationRequest:
    properties:
      distance_to_coast:
        example: 16
        type: number
      elevation:
        example: 12
        type: number
      latitude:
        example: 21.029
        type: number
      longitude:
        example: 105.855
        type: number
      name:
        example: Updated Station Alpha
        type: string
      note:
        example: Updated radar station
        type: string
      status:
        example: INACTIVE
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for the Radar Hub Manager system
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Radar Hub Manager API
  version: "1.0"
paths:
...
  /stations:
    get:
      consumes:
      - application/json
      description: Get a list of all radar stations. Only users with ADMIN role can
        perform this action.
      produces:
      - application/json
      responses:
        "200":
          description: List of stations
          schema:
            items:
              $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Station'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all stations (Admin only)
      tags:
      - stations
    post:
      consumes:
      - application/json
      description: Create a new radar station. Only users with ADMIN role can perform
        this action.
      parameters:
      - description: Station creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateStationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Station created successfully
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Station'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "409":
          description: Conflict - Station already exists
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new station (Admin only)
      tags:
      - stations
  /stations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a radar station. Only users with ADMIN role can perform
        this action.
      parameters:
      - description: Station ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Station deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Station not found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a station (Admin only)
      tags:
      - stations
    get:
      consumes:
      - application/json
      description: Get station information by ID. Only users with ADMIN role can perform
        this action.
      parameters:
      - description: Station ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Station information
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Station'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Station not found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get station by ID (Admin only)
      tags:
      - stations
    put:
      consumes:
      - application/json
      description: Update station information. Only users with ADMIN role can perform
        this action.
      parameters:
      - description: Station ID
        in: path
        name: id
        required: true
        type: integer
      - description: Station update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateStationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Station updated successfully
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Station'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Station not found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing station (Admin only)
      tags:
      - stations
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"